refactor river level display to split functionality to separate files
