+++ src/river_level_display.rs
@@ -1,5 +1,8 @@
 mod css_keyframe_animation;
 mod wave_animation;
+mod waves;
+mod depth_meter;
+mod height_to_screenspace;
 
 use yew::prelude::*;
 
@@ -19,8 +22,8 @@ pub fn river_level_display_background(BackgroundProps {barbourne_last_reading, d
     let wave_animation = use_memo(
         (*barbourne_last_reading, *diglis_last_reading), 
         |(barbourne_reading, diglis_reading)| -> WaveAnimation {
-            let barbourne_flood_percentage = get_flood_percentage(barbourne_reading.value as f64, 2.0, 4.0);
-            let diglis_flood_percentage = get_flood_percentage(diglis_reading.value as f64, 1.5, 3.3);
+            let barbourne_flood_percentage = height_to_screenspace::get_flood_percentage(barbourne_reading.value as f64, 2.0, 4.0);
+            let diglis_flood_percentage = height_to_screenspace::get_flood_percentage(diglis_reading.value as f64, 1.5, 3.3);
             WaveAnimation::new(1.0 - barbourne_flood_percentage, 1.0 - diglis_flood_percentage)
         }
     );
@@ -31,258 +34,13 @@ pub fn river_level_display_background(BackgroundProps {barbourne_last_reading, d
                 class="full_canvas"
                 style={wave_animation.to_string()}
             >
-                <Waves/>
-                <RiverLevelMarkers/>
-                <SolubleText
+                <waves::Waves/>
+                <depth_meter::RiverLevelMarkers/>
+                <waves::SolubleText
                     shown_above_water="No"
                     shown_below_water="Yes"
                 />
             </div>
         </>
     }
-}
-
-#[function_component(Waves)]
-fn waves() -> Html {
-    html! {
-        <>
-            <div
-                class="full_canvas background_water"
-                color="aqua"
-            />
-            <div 
-                class="full_canvas background_water_2"
-                color="aqua"
-            />
-            <div 
-                class="full_canvas clipping_water ShowBelowWaterLevel"
-                color="blue"
-            />
-        </>
-    }
-}
-
-
-#[derive(Properties, PartialEq)]
-pub struct SolubleTextProps {
-    pub shown_above_water: String,
-    pub shown_below_water: String
-}
-
-#[function_component(SolubleText)]
-fn soluble_text(SolubleTextProps {shown_above_water, shown_below_water}: &SolubleTextProps) -> Html {
-    html! {
-        <>
-            <div
-                class="full_canvas result_wrapper ShowAboveWaterLevel"
-            >
-                <h1
-                    class="result_content"
-                    style="color: black;"
-                >
-                    {shown_above_water}
-                </h1>
-            </div>
-            <div
-                class="full_canvas result_wrapper ShowBelowWaterLevel"
-            >
-                <h1
-                    class="result_content" 
-                    style="color: white;"
-                >
-                    {shown_below_water}
-                </h1>
-            </div>
-        </>
-    }
-}
-
-#[function_component(RiverLevelMarkers)]
-fn river_level_markers() -> Html {
-    html! {
-        <div class="depth_meter_container">
-            <div class="left marker">
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"5m"}</div>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"4m"}</div>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"3m"}</div>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"2m"}</div>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"1m"}</div>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block large"/>
-                    <div class="height_label">{"0m"}</div>
-                </div>
-            </div>
-            <div class="right marker">
-                <div class="depth_row">
-                    <div class="height_label">{"5m"}</div>
-                    <div class="block large"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="height_label">{"4m"}</div>
-                    <div class="block large"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="height_label">{"3m"}</div>
-                    <div class="block large"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="height_label">{"2m"}</div>
-                    <div class="block large"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="height_label">{"1m"}</div>
-                    <div class="block large"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="block small"/>
-                </div>
-                <div class="depth_row">
-                    <div class="height_label">{"0m"}</div>
-                    <div class="block large"/>
-                </div>
-            </div>
-        </div>
-    }
-}
-
-fn get_flood_percentage(current_level: f64, normal_level: f64, risky_level: f64) -> f64 {
-	let quarter = risky_level - normal_level; // Calculate what 25% of the range is
-	let base = normal_level - quarter; // Calculate what value 0% would be, aka the base of the range
-
-	let scaled_level = current_level - base;
-	let percentage_multiplier = 1.0 / (4.0 * quarter);
-	let percentage = scaled_level * percentage_multiplier;
-	percentage.min(1.0).max(0.0)
 }
